+++ src/app/Cards/cart-list/cart-item/cart-item.component.ts
@@ -14,6 +14,7 @@ export class CartItemComponent {
   @Output()
   deleteFromCart: EventEmitter<CartModel> = new EventEmitter<CartModel>();
   
+
   onDeleteFromCart(): void {
     console.log(this.product)
     this.deleteFromCart.emit(this.product);