+++ src/app/Cards/cart.module.ts
@@ -3,14 +3,14 @@ import { CommonModule } from '@angular/common';
 import { CartListComponent } from './cart-list/cart-list.component';
 import { CartItemComponent } from './cart-list/cart-item/cart-item.component';
 import { SharedModule } from '../shared/shared.module';
+import { FormsModule } from '@angular/forms';
 
 
 
 @NgModule({
   declarations: [CartListComponent, CartItemComponent],
   imports: [
-    CommonModule,
-    SharedModule
+    SharedModule      
   ],
   exports: [CartListComponent]
 })