+++ src/app/Products/product-list/product-list.component.ts
@@ -5,6 +5,7 @@ import { Category } from 'src/app/models/category';
 import { CartModel } from 'src/app/models/cart.model';
 import { ProductModel } from '../../models/product.model';
 import { ProductsService } from '../products.service';
+import { Observable, of } from 'rxjs';
 
 @Component({
   selector: 'app-product-list',
@@ -12,7 +13,7 @@ import { ProductsService } from '../products.service';
   styleUrls: ['./product-list.component.scss']
 })
 export class ProductListComponent implements OnInit {
-  products: ProductModel[] = []
+  products: Observable<ProductModel[]> = of([]);
   name: string = '';
 
   constructor(private productService: ProductsService) { }
@@ -21,15 +22,6 @@ export class ProductListComponent implements OnInit {
     this.products = this.productService.getProducts();
   }
 
-  filter(): void {
-    if (this.name == '') {
-      this.products = this.productService.getProducts();
-    }
-    else {
-      this.products = this.productService.filterByName(this.name);
-    }
-
-  }
 
   onAddToCart(product: ProductModel): void {
     const cart: CartModel = {