+++ src/app/models/cart.model.ts
@@ -1,9 +1,8 @@
 export interface CartModel {
-    name: string ;
-      description: string ;
-      price: number ;
-      category :string;
-      isAvailable: boolean;
-      quantity:number;
-    }
-    
+  name: string;
+  description: string;
+  price: number;
+  category: string;
+  isAvailable: boolean;
+  quantity: number;
+}